+++ test/harness/services/gossip/adapter/transport_contract_test.go
@@ -62,7 +62,7 @@ func createContractTestConfig(publicKey primitives.Ed25519PublicKey, gossipListe
 	cfg := config.EmptyConfig()
 	cfg.SetNodePublicKey(publicKey)
 	cfg.SetGossipPeers(gossipPeers)
-	cfg.SetUint16(config.GOSSIP_LISTEN_PORT, gossipListenPort)
+	cfg.SetUint32(config.GOSSIP_LISTEN_PORT, uint32(gossipListenPort))
 	cfg.SetDuration(config.GOSSIP_CONNECTION_KEEP_ALIVE_INTERVAL, 20*time.Millisecond)
 	cfg.SetDuration(config.GOSSIP_NETWORK_TIMEOUT, 1*time.Second)
 	return cfg
